requirements.txt
cassandra-driver
requests
random
flask